backend-python/pyproject.toml
[project]
name = "backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dependency-injector>=4.48.3",
    "google-genai>=1.56.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "httpx>=0.28.1",
    "mypy-protobuf>=3.7.0",
    "openai>=2.14.0",
    "protobuf>=6.33.2",
    "pydantic-settings>=2.12.0",
    "types-protobuf>=6.32.1.20251210",
]

[dependency-groups]
dev = ["pytest>=9.0.2", "pytest-asyncio>=1.3.0", "pytest-mock>=3.15.1"]
